requirements.txt
python-dotenv==1.1.0
schedule==1.2.2
